backend/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# PDF processing
PyMuPDF>=1.23.8

# Vector database
faiss-cpu>=1.8.0

# OpenAI (embeddings + LLM)
openai>=1.12.0

# Data validation
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.26.0
